Tests/LibWeb/Text/expected/HTML/svg-script-parser-reentrant-document-write.txt

svg write parent: body
body children: svg,span,script
